+++ serving-api/src/main/scala/com/stratio/sparta/serving/api/actor/NginxActor.scala
@@ -9,27 +9,34 @@ import java.util.concurrent.TimeUnit
 
 import akka.actor.{Actor, Cancellable}
 import akka.stream.ActorMaterializer
+
 import com.stratio.sparta.serving.api.actor.NginxActor._
 import com.stratio.sparta.serving.core.actor.ExecutionStatusChangePublisherActor.ExecutionStatusChange
 import com.stratio.sparta.serving.core.models.enumerators.WorkflowStatusEnum._
 import com.stratio.sparta.serving.core.models.workflow.WorkflowExecution
 import com.stratio.sparta.serving.core.utils.NginxUtils
 import com.stratio.sparta.serving.core.utils.NginxUtils._
-
 import scala.concurrent.Future
 import scala.concurrent.duration._
+
+import akka.cluster.pubsub.DistributedPubSub
+import akka.cluster.pubsub.DistributedPubSubMediator.{Subscribe, Unsubscribe}
+
+import com.stratio.sparta.serving.core.actor.ExecutionStatusChangePublisherActor
+
 
 
 class NginxActor extends Actor {
 
   import context.dispatcher
+  val mediator = DistributedPubSub(context.system).mediator
 
   private lazy val scheduledInitialCheck: Cancellable =
     context.system.scheduler.scheduleOnce(30.seconds, self, UpdateAndMakeSureIsRunning)
 
 
   override def preStart(): Unit = {
-    context.system.eventStream.subscribe(self, classOf[ExecutionStatusChange])
+    mediator ! Subscribe(ExecutionStatusChangePublisherActor.ClusterTopicExecutionStatus, self)
     scheduledInitialCheck
   }
 
@@ -85,10 +92,10 @@ class NginxActor extends Actor {
   }
 
   override def postStop(): Unit = {
+    mediator ! Unsubscribe(ExecutionStatusChangePublisherActor.ClusterTopicExecutionStatus, self)
     scheduledPeriodicCheck.cancel()
     scheduledInitialCheck.cancel()
     updateAndMakeSureIsRunning
-    context.system.eventStream.unsubscribe(self, classOf[ExecutionStatusChange])
   }
 }