+++ serving-api/src/main/scala/com/stratio/sparta/serving/api/helpers/SpartaHelper.scala
@@ -6,12 +6,10 @@
 
 package com.stratio.sparta.serving.api.helpers
 
-import scala.util.{Properties, Try}
 import akka.actor.{ActorSystem, Props}
+import akka.cluster.Cluster
 import akka.event.slf4j.SLF4JLogging
 import akka.io.IO
-import org.apache.ignite.Ignition
-import spray.can.Http
 import com.stratio.sparta.serving.api.actor._
 import com.stratio.sparta.serving.api.service.ssl.SSLSupport
 import com.stratio.sparta.serving.core.actor._
@@ -22,6 +20,11 @@ import com.stratio.sparta.serving.core.factory.PostgresFactory
 import com.stratio.sparta.serving.core.helpers.SecurityManagerHelper
 import com.stratio.sparta.serving.core.services.migration.CassiopeiaMigrationService
 import com.stratio.sparta.serving.core.utils.SpartaIgnite
+import com.typesafe.config.ConfigFactory
+import org.apache.ignite.Ignition
+import spray.can.Http
+
+import scala.util.{Properties, Try}
 
 
 /**
@@ -56,36 +59,40 @@ object SpartaHelper extends SLF4JLogging with SSLSupport {
       SecurityManagerHelper.initCrossdataSecurityManager()
 
       log.debug("Initializing Sparta system ...")
-      implicit val system = ActorSystem(appName, SpartaConfig.getSpartaConfig())
+      implicit val system = ActorSystem(appName, SpartaConfig.getSpartaConfig().get.withFallback(ConfigFactory.load().getConfig("clusterSparta")))
+      system.actorOf(Props[SpartaClusterNodeActor], "clusterNode")
+      Cluster(system) registerOnMemberUp {
 
-      val parametersListenerActor = system.actorOf(Props[ParametersListenerActor])
-      val executionStatusChangeListenerActor = system.actorOf(Props[ExecutionStatusChangeListenerActor])
+        val parametersListenerActor = system.actorOf(Props[ParametersListenerActor])
+        val executionStatusChangeListenerActor = system.actorOf(Props(new ExecutionStatusChangeListenerActor()))
 
-      system.actorOf(Props[SchedulerMonitorActor])
-      system.actorOf(Props(new ExecutionStatusChangePublisherActor()))
+        system.actorOf(Props[SchedulerMonitorActor])
+        system.actorOf(Props(new ExecutionStatusChangePublisherActor()))
 
-      if (Try(SpartaConfig.getDetailConfig().get.getBoolean("lineage.enable")).getOrElse(false)) {
-        log.info("Initializing lineage service ...")
-        //TODO lineage
-        //system.actorOf(LineageService.props(executionListenerActor, workflowListenerActor))
+        if (Try(SpartaConfig.getDetailConfig().get.getBoolean("lineage.enable")).getOrElse(false)) {
+          log.info("Initializing lineage service ...")
+          //TODO lineage
+          //system.actorOf(LineageService.props(executionListenerActor, workflowListenerActor))
+        }
+
+        val controllerActor = system.actorOf(Props(new ControllerActor(
+          executionStatusChangeListenerActor,
+          parametersListenerActor
+        )), ControllerActorName)
+
+        log.info("Binding Sparta API ...")
+        IO(Http) ! Http.Bind(controllerActor,
+          interface = SpartaConfig.getApiConfig().get.getString("host"),
+          port = SpartaConfig.getApiConfig().get.getInt("port")
+        )
+
+        if (Properties.envOrNone(NginxMarathonLBHostEnv).fold(false) { _ => true })
+          Option(system.actorOf(Props(new NginxActor()), NginxActorName))
+
+        log.info("Sparta server initiated successfully")
       }
-
-      val controllerActor = system.actorOf(Props(new ControllerActor(
-        executionStatusChangeListenerActor,
-        parametersListenerActor
-      )), ControllerActorName)
-
-      log.info("Binding Sparta API ...")
-      IO(Http) ! Http.Bind(controllerActor,
-        interface = SpartaConfig.getApiConfig().get.getString("host"),
-        port = SpartaConfig.getApiConfig().get.getInt("port")
-      )
-
-      if (Properties.envOrNone(NginxMarathonLBHostEnv).fold(false) { _ => true })
-        Option(system.actorOf(Props(new NginxActor()), NginxActorName))
-
-      log.info("Sparta server initiated successfully")
-    } else log.info("Sparta configuration is not defined")
+    }
+    else log.info("Sparta configuration is not defined")
 
     Runtime.getRuntime.addShutdownHook(new Thread() {
       override def run(): Unit = {