+++ serving-core/src/main/scala/com/stratio/sparta/serving/core/actor/ListenerPublisher.scala
@@ -7,13 +7,22 @@
 package com.stratio.sparta.serving.core.actor
 
 import akka.actor.Actor
+import akka.cluster.Cluster
+import akka.cluster.pubsub.DistributedPubSub
+import akka.cluster.pubsub.DistributedPubSubMediator.{Subscribe, Unsubscribe}
 import akka.event.slf4j.SLF4JLogging
 import org.apache.curator.framework.recipes.cache.PathChildrenCache
 
 import com.stratio.sparta.serving.core.actor.BusNotification.InitNodeListener
 import com.stratio.sparta.serving.core.models.SpartaSerializer
+import com.stratio.sparta.serving.core.utils.SpartaClusterUtils
 
-trait ListenerPublisher extends Actor with SpartaSerializer with SLF4JLogging {
+trait ListenerPublisher extends Actor with SpartaSerializer with SLF4JLogging  with SpartaClusterUtils {
+
+  val cluster = Cluster(context.system)
+  val mediator = DistributedPubSub(context.system).mediator
+
+  import ListenerPublisher._
 
   protected var pathCache: Option[PathChildrenCache] = None
 
@@ -24,8 +33,35 @@ trait ListenerPublisher extends Actor with SpartaSerializer with SLF4JLogging {
       log.info(s"Restarting zkNode Listeners for ${self.toString()}")
       initNodeListener()
     }
-    case _ =>
-      log.debug(s"Unrecognized message in ${self.toString()}")
+  }
+
+  def initNodeListener(): Unit
+
+  override def preStart(): Unit = {
+    mediator ! Subscribe(ClusterTopicNodeListener, self)
+    initNodeListener()
+  }
+
+  override def postStop(): Unit = {
+    mediator ! Unsubscribe(ClusterTopicNodeListener, self)
+    pathCache.foreach(_.close())
+  }
+
+  def isClusterLeader(): Boolean = isThisNodeClusterLeader(cluster)
+}
+
+
+trait MarathonListenerPublisher extends Actor with SpartaSerializer with SLF4JLogging {
+
+  protected var pathCache: Option[PathChildrenCache] = None
+
+  val relativePath : String
+
+  def listenerReceive: Receive = {
+    case cd: InitNodeListener => {
+      log.info(s"Restarting zkNode Listeners for ${self.toString()}")
+      initNodeListener()
+    }
   }
 
   def initNodeListener(): Unit
@@ -39,4 +75,9 @@ trait ListenerPublisher extends Actor with SpartaSerializer with SLF4JLogging {
     context.system.eventStream.unsubscribe(self, classOf[InitNodeListener])
     pathCache.foreach(_.close())
   }
+}
+
+object ListenerPublisher {
+
+  val ClusterTopicNodeListener = "NodeListener"
 }